requirements.txt
# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# AWS
boto3>=1.35.0

# Document processing
pypdf>=5.0.0
markdown>=3.6.0

# Observability
structlog>=24.4.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp>=1.28.0
opentelemetry-exporter-otlp-proto-http>=1.28.0
opentelemetry-instrumentation-fastapi>=0.49b0

# HTTP client (tests, async)
httpx>=0.27.0

# Config
pydantic-settings>=2.6.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
